+++ udata/core/site/api.py
@@ -6,7 +6,6 @@ from udata.api import API, api, fields
 from udata.auth import admin_permission
 from udata.core.dataservices.models import Dataservice
 from udata.core.dataset.api_fields import dataset_fields
-from udata.core.reuse.api_fields import reuse_fields
 from udata.models import Dataset, Reuse
 from udata.rdf import CONTEXT, RDF_EXTENSIONS, graph_response, negociate_content
 from udata.utils import multi_to_dict
@@ -62,7 +61,7 @@ class SiteHomeDatasetsAPI(API):
 class SiteHomeReusesAPI(API):
     @api.doc("get_home_reuses")
     @api.secure(admin_permission)
-    @api.marshal_list_with(reuse_fields)
+    @api.marshal_list_with(Reuse.__read_fields__)
     def get(self):
         """List homepage featured reuses"""
         return current_site.settings.home_reuses
@@ -70,7 +69,7 @@ class SiteHomeReusesAPI(API):
     @api.secure(admin_permission)
     @api.doc("set_home_reuses")
     @api.expect(([str], "Reuse IDs to put in homepage"))
-    @api.marshal_list_with(reuse_fields)
+    @api.marshal_list_with(Reuse.__read_fields__)
     def put(self):
         """Set the homepage reuses editorial selection"""
         if not isinstance(request.json, list):