+++ udata/tests/reuse/test_reuse_model.py
@@ -1,4 +1,3 @@
-import re
 from datetime import datetime
 
 from udata.core.dataset import tasks as dataset_tasks