+++ udata/tests/test_cors.py
@@ -1,11 +1,8 @@
 from datetime import datetime
 
-import pytest
 from flask import url_for
 
 from udata.core.dataset.factories import DatasetFactory
-from udata.core.topic.factories import TopicFactory
-from udata.search import reindex
 from udata.tests.api import APITestCase
 from udata.tests.helpers import assert_status